+++ spec/support/helper_methods.rb
@@ -18,15 +18,14 @@ end
 def add_signatures(hash, klass=described_class)
   properties = klass.new(hash).send(:xml_elements)
   hash[:author_signature] = properties[:author_signature]
-  hash[:parent_author_signature] = properties[:parent_author_signature]
 end
 
 def sign_with_key(privkey, signature_data)
-  Base64.strict_encode64(privkey.sign(OpenSSL::Digest::SHA256.new, signature_data))
+  Base64.strict_encode64(privkey.sign(OpenSSL::Digest.new("SHA256"), signature_data))
 end
 
 def verify_signature(pubkey, signature, signed_string)
-  pubkey.verify(OpenSSL::Digest::SHA256.new, Base64.decode64(signature), signed_string)
+  pubkey.verify(OpenSSL::Digest.new("SHA256"), Base64.decode64(signature), signed_string)
 end
 
 # time helper